+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Selected tab color in dark mode
+- Breadcrumb links
 
 ## [v1.3.0] - 2025-01-09